app/src/requirements.txt
redis
pandas
Flask
line-bot-sdk
MySQLdb
Flask-API
mecab-python3